+++ src/Plugin/ModelMethodGenerator/Image.php
@@ -2,12 +2,18 @@
 
 namespace Drupal\wmscaffold\Plugin\ModelMethodGenerator;
 
+use Drupal\image\Plugin\Field\FieldType\ImageItem;
+
 /**
  * @ModelMethodGenerator(
  *     id = "image",
  *     provider = "image",
  * )
  */
-class Image extends EntityReference
+class Image extends BaseFieldItem
 {
+    public static function getType(): ?string
+    {
+        return ImageItem::class;
+    }
 }