+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix optional datetime getters
 - Validate whether modules are installed for commands interacting with contrib
+- Fix image getters not returning an ImageItem instance
 
 ## [1.13.2] - 2022-01-29
 ### Fixed